+++ Capri.Services/Account/ILoginService.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Capri.Web.ViewModels.User;
 
-namespace Capri.Services
+namespace Capri.Services.Account
 {
     public interface ILoginService
     {
