--- Capri.Services/ITokenGenerator.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Capri.Database.Entities.Identity;
-
-namespace Capri.Services
-{
-    public interface ITokenGenerator
-    {
-        string GenerateTokenFor(User user);
-    }
-}
