+++ Capri.Services/Promoters/IPromoterCreator.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Capri.Database.Entities;
 using Capri.Web.ViewModels.Promoter;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public interface IPromoterCreator
     {
