+++ Capri.Services/Promoters/IPromoterDeleter.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Capri.Database.Entities;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public interface IPromoterDeleter
     {
