+++ Capri.Services/Promoters/IPromoterGetter.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Capri.Database.Entities;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public interface IPromoterGetter
     {
