+++ Capri.Services/Promoters/IPromoterUpdater.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Capri.Database.Entities;
 using Capri.Web.ViewModels.Promoter;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public interface IPromoterUpdater
     {
