+++ Capri.Services/Promoters/PromoterCreator.cs
@@ -1,13 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Linq;
-using System.Text;
 using Capri.Database;
 using Capri.Database.Entities;
+using Capri.Services.Users;
 using Capri.Web.ViewModels.Promoter;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public class PromoterCreator : IPromoterCreator
     {
