+++ Capri.Services/Promoters/PromoterGetter.cs
@@ -1,13 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Linq;
-using System.Text;
 using Microsoft.EntityFrameworkCore;
 using Capri.Database;
 using Capri.Database.Entities;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public class PromoterGetter : IPromoterGetter
     {
