+++ Capri.Services/Promoters/PromoterUpdater.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Linq;
-using System.Text;
 using Microsoft.EntityFrameworkCore;
 using AutoMapper;
 using Capri.Database;
 using Capri.Database.Entities;
+using Capri.Services.Users;
 using Capri.Web.ViewModels.Promoter;
 using Capri.Web.ViewModels.User;
 
-namespace Capri.Services
+namespace Capri.Services.Promoters
 {
     public class PromoterUpdater : IPromoterUpdater
     {
