+++ Capri.Services/Proposals/IProposalCreator.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Capri.Database.Entities;
-using Capri.Database.Entities.Identity;
 using Capri.Web.ViewModels.Proposal;
 
 namespace Capri.Services.Proposals
