+++ Capri.Services/Token/ITokenGenerator.cs
@@ -1,0 +1,9 @@
+﻿using Capri.Database.Entities.Identity;
+
+namespace Capri.Services.Token
+{
+    public interface ITokenGenerator
+    {
+        string GenerateTokenFor(User user);
+    }
+}
