+++ Capri.Services/Users/IUserCreator.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Capri.Database.Entities.Identity;
 
-namespace Capri.Services
+namespace Capri.Services.Users
 {
     public interface IUserCreator
     {
