+++ Capri.Services/Users/IUserGetter.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Capri.Database.Entities.Identity;
+
+namespace Capri.Services.Users
+{
+    public interface IUserGetter
+    {
+        Task<User> GetCurrentUser();
+    }
+}
