+++ Capri.Services/Users/IUserUpdater.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Capri.Database.Entities.Identity;
 using Capri.Web.ViewModels.User;
 
-namespace Capri.Services
+namespace Capri.Services.Users
 {
     public interface IUserUpdater
     {
