+++ Capri.Services/Users/UserUpdater.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
@@ -8,7 +6,7 @@ using Capri.Database;
 using Capri.Database.Entities.Identity;
 using Capri.Web.ViewModels.User;
 
-namespace Capri.Services
+namespace Capri.Services.Users
 {
     public class UserUpdater : IUserUpdater
     {
