+++ Capri.Web/Configuration/MapperConfiguration.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using Capri.Services;
+﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace Capri.Web.Configuration
 {
